Switch to YouTube React component
Installs the 'react-youtube' component [1], which greatly simplifies the
use of the YouTube embed API and iframe.

Also simplifies the CSS style to remove the use of flexbox and forces
the YouTube player to fill the entire viewport height.

Also adds a default constructor to fix an error thrown when loading the
game for the first time.

[1] https://www.npmjs.com/package/react-youtube