Fix infinite play-pause loop; support auto-play
Stores the ID of the user who pressed play or pause in Firebase so we
can check it in our data change handler. If the changed data has a user
ID that matches ours, we obviously made that change and we should ignore
this event.

Maintains a React state mirror of the all Firebase data, so we can
always access the latest data in using this.state. We use this to add
checks to the on-play and on-pause callbacks, where we only write to
Firebase if it would results in a state change.

For example, if a Firebase change resulted in my client to start playing
a video (i.e. I didn't click "play"), the video would auto-start and
the on-play callback fires. We do *not* want to update Firebase in this
case, so we check if the current Firebase state says a video is playing
(and do nothing in this case).

Also adds auto-play to the on-ready callback.